Started restructuring the compiler to use many passes, incrementally simplifying the AST.
In particular, I wanted to add a pass that replaces named variables with "Load(address)" expressions; this eliminates the need to deal with scopes and variable names in later code.

Not all the tests pass, but fixing them to work with the new multi-pass system will require too many changes to make in one commit.